requirements.txt
fastapi~=0.109.2
mistralai~=0.0.11
openai~=1.8.0
openpyxl
pandas~=2.1.4
pydantic~=2.5.3
python-dotenv~=1.0.1
python-multipart
requests~=2.31.0
scikit-learn~=1.4.0
tiktoken~=0.5.2
tqdm~=4.66.1
uvicorn
rdkit~=2023.9.6
scipy~=1.11.4
drfp~=0.3.6
anthropic~=0.25.3
numpy~=1.26.3
gigachat~=0.1.27